KUDU-3495 Initialize the bitmap while decoding insert/upsert
The upsert request with an old client schema might lead to
unexpected result eventually.

If client A opens a table first and then client B add a column
to the table, the upsert request of client A might set the
newly added column to default value.

The reason is that we did not initialize the tablet_isset_bitmap
while decoding the insert or upsert requests.

A user does not want to change the column which he did not request
to even if the client schema is not correct, so initialize the bitmap.

Change-Id: I38a33cb58a085bc83854c5145e904c8ed51092a8
Reviewed-on: http://gerrit.cloudera.org:8080/20327
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>
(cherry picked from commit f1f3c907fc6f012b6d7a72f0103e1db4f56f0e7d)
Reviewed-on: http://gerrit.cloudera.org:8080/21296
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>
Tested-by: Alexey Serbin <alexey@apache.org>