KUDU-3562 fix integer overflow in available space metrics
This patch addresses KUDU-3562.  The problem was in implicit casting
of 64-bit integers down to 32-bit ones when introducing a helper
variable in CalculateAvailableSpace().  A new test scenario is added
to cover the fixed issue and catch regressions in future, if any.

In addition, this patch corrects the type of the related metrics in
disk_failure-itest.cc.  It also clarifies on the description and tags
of the --fs_{data_dirs,wal_dir}_available_space_cache_seconds flags.

Change-Id: I974aea822626e4648886388c0de3741ac459f2ec
Reviewed-on: http://gerrit.cloudera.org:8080/21227
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Abhishek Chennaka <achennaka@cloudera.com>
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
(cherry picked from commit e54c80cf9ba129b943eed6af3266899753b1b4ac)
Reviewed-on: http://gerrit.cloudera.org:8080/21228