Update build pattern for fetching flaky tests list
During the infra changes which happened in the last months, the build_id
prefix has been changed [note 1]. As a result flaky tests are not
retried during testing in the pre-commit runs. This patch updates the
build pattern in the build-and-test.sh script, such that the
list of flaky tests is fetched correctly.

[note 1]:
The build_id of an isolated test is constructed with Jenkins job's name
and build number, prefixed with "jenkins-". To illustrate this let me
give some examples, for build_ids prior and after the infra upgrade.
Before:
jenkins-kudu-gerrit-BUILD_TYPE=ASAN-29232
jenkins-kudu-gerrit-BUILD_TYPE=DEBUG-29227
After:
jenkins-build_and_test-64
jenkins-build_and_test-63
From the above we can see that the new Jenkins job name caused the
issue. The new job is parametrised, does not contain the build type in
the job's name. Therefore, a change in the naming is justified.
In case something like this happens in the future, the build_ids can be
observed in the test_results MySQL database on the isolateserver test
infra host.

Change-Id: I317a3a32c06c06306b97566f954e0ffd508ce01f
Reviewed-on: http://gerrit.cloudera.org:8080/21327
Reviewed-by: Ashwani Raina <araina@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>
(cherry picked from commit 7b5fc9f2f0ad538242656431bf76e253cc537772)
Reviewed-on: http://gerrit.cloudera.org:8080/21466
Reviewed-by: Marton Greber <greber.mx@gmail.com>
Reviewed-by: Zoltan Chovan <zchovan@cloudera.com>
Tested-by: Attila Bukor <abukor@apache.org>