KUDU-3661 Ranger policy not honored in Kudu
This fixes a long-standing bug in the Ranger authorization provider
where we return prematurely from
RangerAuthzProvider::FillTablePrivilegePB() when the SELECT
action is encountered while iterating through
unordered_set<ActionPB, ActionHas> container, potentially
resulting in missing privileges depending on the position of the
SELECT action in the set. While this behavior depends on the
libc++/libstdc++ implementation, we have observed reports of this
issue on RHEL/CentOS 8 machines.

Thanks to Alexey Serbin for valuable inputs and contributions on
this fix.
Test case source: https://gerrit.cloudera.org/#/c/22809/

Change-Id: I635132154d622eb41e993a0a1a818b21b5af6bb7
Reviewed-on: http://gerrit.cloudera.org:8080/22806
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Abhishek Chennaka <achennaka@cloudera.com>