Fix compilation if using OpenSSL < 1.1.1
This change adds a version guard around the call to
SSL_CTX_set_ciphersuites() so it only compiles when
building with OpenSSL versions that support it. This
addresses a build failure caused by commit f6fbcde on
systems using OpenSSL 1.0.x.

The --webserver_tls_min_protocol flag's validator is
also updated to check for the set TLS version against
the available version.

The build has been tested successfully in CentOS 7.9
machine.

This is a followup to f6fbcde.

Change-Id: I7410a1f08e700efd8a917be8c59d54a497023cef
Reviewed-on: http://gerrit.cloudera.org:8080/23295
Tested-by: Abhishek Chennaka <achennaka@cloudera.com>
Reviewed-by: Alexey Serbin <alexey@apache.org>