KUDU-1261 add GetArray/SetArray C++ client API
This patch adds methods into the Kudu client C++ API to set an array
of values for all the supported scalar types except for
INT128 and DECIMAL128.

Also, added corresponding KuduScanBatch::RowPtr::GetArrayXxx() methods.

Added src/kudu/client/array_cell.h with KuduArrayCellView API
to access raw data in array cells.  That's oriented to working with
raw data access API such as KuduScanBatch::RowPtr::cell() and
KuduScanBatch::direct_data(); AFAIK, that's what Impala uses to fetch
scan results from Kudu.

This patch is made to unblock the integration work dependent on the new
Kudu client API for array columns (i.e., integration with Impala).
So, it contains only partial test coverage for the newly added
functionality, primarily in partial_row-test.cc.  The end-to-end tests
for writing/reading array data are coming in a follow-up changelist.
They require server-side support of the array data in both MRS and
array data blocks persisted on the disk, and is a part of another
changelist [1].

Adding Doxygen documentation comments for the newly introduced methods
will be taken care in a separate changelist as well.

[1] https://gerrit.cloudera.org/#/c/22868/

Change-Id: Idba29337563cefefa8cf1000663ec7e27885cb6f
Reviewed-on: http://gerrit.cloudera.org:8080/23474
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>